move compress request body test cases to own file
