+++ tests/cases/mod.rs
@@ -1,2 +1,3 @@
+mod compress_request_body;
 mod download;
 mod logging;
